pyproject.toml
[project]
name = "openmate-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "sunq99",email = "sunq99@openmate.co.kr"}
]
requires-python = ">=3.11,<3.15"
dependencies = [
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "langchain (>=1.2.9,<2.0.0)",
    "langchain-google-genai (>=4.2.0,<5.0.0)",
    "pandas (>=1.4.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn (>=0.34.0,<1.0.0)",
    "streamlit (>=1.54.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
